requirements/base.txt
alembic==1.11.1
blinker==1.6.2
Cerberus==1.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-JWT-Extended==4.5.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
py-bcrypt==0.4
PyJWT==2.7.0
requests==2.31.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
tabulate==0.9.0
toml==0.10.2
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
